Edit code to CodingStandard
