+++ src/main/java/command/AddDeadlineCommand.java
@@ -1,10 +1,10 @@
-package Command;
+package command;
 
-import Amogus.AmogusException;
-import Amogus.FileStorage;
-import Amogus.UI;
-import Tasks.Deadlines;
-import Tasks.TaskList;
+import amogus.AmogusException;
+import amogus.FileStorage;
+import amogus.UI;
+import tasks.Deadlines;
+import tasks.TaskList;
 
 public class AddDeadlineCommand implements Command {
