+++ src/main/java/command/Command.java
@@ -1,9 +1,9 @@
-package Command;
+package command;
 
-import Amogus.AmogusException;
-import Amogus.FileStorage;
-import Amogus.UI;
-import Tasks.TaskList;
+import amogus.AmogusException;
+import amogus.FileStorage;
+import amogus.UI;
+import tasks.TaskList;
 
 public interface Command {
