+++ src/main/java/tasks/Task.java
@@ -1,8 +1,9 @@
-package Tasks;
+package tasks;
 
 public class Task {
-    protected String description;
-    protected boolean isDone;
+
+    private String description;
+    private boolean isDone;
 
     public Task(String description) {
         this.description = description;
