+++ src/main/java/tasks/TaskList.java
@@ -1,4 +1,4 @@
-package Tasks;
+package tasks;
 
 import java.util.ArrayList;
