+++ src/test/java/tasks/TaskListTest.java
@@ -1,4 +1,4 @@
-package Tasks;
+package tasks;
 
 import org.junit.jupiter.api.Test;
